Exam Papar Ganerator/requirements.txt
Flask==2.0.1
python-dotenv==1.0.0
Werkzeug==3.0.1
google-generativeai==0.3.2 